A few changes
'Izvod vijesti' is not a particularly good translation, and it would confuse a lot of people, so I believe it's better to use 'ATOM novosti' instead. Nobody uses 'izvod' for ATOM feeds.

Also, 'Označi kao omiljeno' is only used for marking posts as favourites, not liking, so I used 'Sviđa mi se' which is a better translation for 'Like', and it's used by Facebook.

I've changed 'pogrešni' to it's synonym 'krivi' because there our linguist are arguing about the spelling of it for years now. Is it 'pogrešni', 'pogrješni' or 'pofsjgnh', nobody knows for sure.

Croatian does not transcribe words fonetically like Serbian, so it isn't 'Tviter', it's 'Twitter'.

I also changed some bad constructions. Of course, zdroid is not to blame, as he's not really a native speaker of Croatian, which is a language that still doesn't have definitive grammar and spelling rules.